[Security Solution] unskip detection_page_filters Cypress test
